+++ src/main/java/kitchenpos/menu/dto/MenuCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.ui.dto;
+package kitchenpos.menu.dto;
 
 import kitchenpos.product.domain.Price;
