+++ src/main/java/kitchenpos/ordertable/application/OrderTableService.java
@@ -3,9 +3,9 @@ package kitchenpos.ordertable.application;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableChangeEmptyEvent;
 import kitchenpos.ordertable.domain.repository.OrderTableRepository;
-import kitchenpos.ordertable.ui.dto.OrderTableCreateRequest;
-import kitchenpos.ordertable.ui.dto.OrderTableResponse;
-import kitchenpos.ordertable.ui.dto.OrderTableUpdateRequest;
+import kitchenpos.ordertable.dto.OrderTableCreateRequest;
+import kitchenpos.ordertable.dto.OrderTableResponse;
+import kitchenpos.ordertable.dto.OrderTableUpdateRequest;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
