+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -5,9 +5,9 @@ import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.product.domain.Price;
 import kitchenpos.product.domain.Product;
-import kitchenpos.menu.ui.dto.MenuCreateRequest;
-import kitchenpos.menu.ui.dto.MenuProductRequest;
-import kitchenpos.menu.ui.dto.MenuResponse;
+import kitchenpos.menu.dto.MenuCreateRequest;
+import kitchenpos.menu.dto.MenuProductRequest;
+import kitchenpos.menu.dto.MenuResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
