+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -5,10 +5,10 @@ import kitchenpos.menu.domain.Menu;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.ui.dto.OrderCreateRequest;
-import kitchenpos.order.ui.dto.OrderLineItemCreateRequest;
-import kitchenpos.order.ui.dto.OrderResponse;
-import kitchenpos.order.ui.dto.OrderUpdateRequest;
+import kitchenpos.order.dto.OrderCreateRequest;
+import kitchenpos.order.dto.OrderLineItemCreateRequest;
+import kitchenpos.order.dto.OrderResponse;
+import kitchenpos.order.dto.OrderUpdateRequest;
 import kitchenpos.ordertable.domain.OrderTable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
