+++ src/test/java/kitchenpos/application/OrderTableServiceTest.java
@@ -4,13 +4,13 @@ import kitchenpos.EntityFactory;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.ui.dto.OrderUpdateRequest;
+import kitchenpos.order.dto.OrderUpdateRequest;
 import kitchenpos.ordertable.application.OrderTableService;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableChangeEmptyEvent;
-import kitchenpos.ordertable.ui.dto.OrderTableCreateRequest;
-import kitchenpos.ordertable.ui.dto.OrderTableResponse;
-import kitchenpos.ordertable.ui.dto.OrderTableUpdateRequest;
+import kitchenpos.ordertable.dto.OrderTableCreateRequest;
+import kitchenpos.ordertable.dto.OrderTableResponse;
+import kitchenpos.ordertable.dto.OrderTableUpdateRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
