+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -4,14 +4,14 @@ import kitchenpos.EntityFactory;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.ui.dto.OrderUpdateRequest;
+import kitchenpos.order.dto.OrderUpdateRequest;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.repository.OrderTableRepository;
-import kitchenpos.product.ui.dto.TableGroupCreateRequest;
-import kitchenpos.product.ui.dto.TableGroupOrderTableRequest;
-import kitchenpos.product.ui.dto.TableGroupResponse;
 import kitchenpos.tablegroup.application.TableGroupService;
 import kitchenpos.tablegroup.domain.TableGroup;
+import kitchenpos.tablegroup.dto.TableGroupCreateRequest;
+import kitchenpos.tablegroup.dto.TableGroupOrderTableRequest;
+import kitchenpos.tablegroup.dto.TableGroupResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
